Fix WARP destroying DMND
If WARP moved in its update function, then the x/y variables in the main
loop were out of date. When the normal movement code ran, changes were
made to the wrong place in the pmap. This resulted in some particles
with low IDs being destroyed by the addition of WARP+SING.

This change might also fix the WARP+PRTI crash that I never managed to
reproduce.