Make ambient heat convection work a bit better
Previously, it was comparing the temperature of the current cell to the
average of the cells around it (plus some advection affecting the
average, but that makes no difference if the air starts out stationary),
and basing the velocity change on this temperature difference.

If the cell below is hotter and the cell above is cooler, the air should
rise. But in this case, the average of surrounding cells tends to be
near the temperature of the current cell, so not much happens.

Just using the temperature difference between the current cell and the
cell above makes convection work a lot better.