Float rounding strikes again - set destination coords using integers when moving particles with PSTN
instead of adding a delta value to the current position, which might not give the correct result. Particles (except solids) were on rare occasions ending up at a point 1 pixel away from where they should be after being pushed by PSTN. This led to stacking, and in the case of save 1732622 after changing BIZS to a liquid, to disintegration of the save.

(TPT++ version of commit 2ad996dfe621887355f8532f1262c306421bc2de in jacksonmj fork)