Make RFRG store old pressure in a less hideous way
I've actually tested this and it looks more promising than the float version ever has been. Pressurising RFRG with temp=273 from p=0 to p=256 heats it up to temp=546±10. While not significantly superior to the float approach, it is better and doesn't cast pointers back and forth between int and float.