Mark saves as being from the next version in dev builds
That is, undo the decoupling of the effective save version from what is today the upstream display version done in 7fc3fb4c15e2. This enables taking advantage of new features, for example the comprehensive palette, in saves that do come with a comprehensive palette, which is a 98.0-and-above feature.

Also fully populate GameSave::version before making checks against it and fix GameSave::fromNewerVersion reflecting the relationship between only the major components of version numbers.