sim.loadStamp: default includePressure to 1
It doesn't make much sense for a Lua function to vary depending on whether the user holds shift or not.

This restores the original behavior of includePressure in 6931f3552018, which was accidentally removed later in a407aba087a9. Since I doubt anyone used this parameter, I'll just keep it at position 6 instead of moving it back to 4.